+++ app/app.js
@@ -20,7 +20,6 @@ import './public/styles/global.scss';
 
 import { Switch, Route, BrowserRouter } from 'react-router-dom';
 
-
 import Header from './components/Header';
 import Footer from './components/Footer';
 
@@ -71,7 +70,6 @@ const render = () => {
             <meta name="description" content="Seal Stats - Hockey Stats" />
           </Helmet>
           <Header />
-
           <Switch>
             <Route exact path="/" component={HomePage} />
             <Route path="/test" component={TestPage} />
