+++ config/nginx.conf.erb
@@ -41,7 +41,7 @@ http {
       try_files $uri $uri/ /index.html;
     }
     location /images/ {
-      root /app/app;
+      root /app/app/public;
     }
     location /graphql {
       index index.html;
